Design flaw reworked
I realized the bottom table of 'add programs' is the csv preview, so i'm going to need to revert that page. Now getting state to programs.
